Fix container build failures
Several containers were failing to build when resources were added to the
pixl_dcmd package.  Fixed by calculating the dependencies with pip-tools
then installing that requirements.txt instead of installing once without
the src directory and again with it.  Removed core as a specified dependency
of pixl_imaging because it breaks calculation of the dependency graph. That's
technically incorrect but we always install it manually anyway.