Don't set a global Hasher instance
This was a bit of a pain to monkeypatch in the tests, as the call to
`Hasher()` here kept trying to initiate an actual AzureKeyVault
instance, which would fail because the proper environment variables are
missing. Doing this at least allows us to monkeypatch the `Hasher` class
in the tests so that it uses a mock keyvault instance.